+++ modules/client-common/src/internal/types.ts
@@ -3,35 +3,46 @@ import { JsonRpcProvider, Network, Networkish } from "@ethersproject/providers";
 import { Client as IpfsClient } from "@aragon/sdk-ipfs";
 import { GraphQLClient } from "graphql-request";
 
-export type NetworkDeployment = {
-  daoFactory: string;
-  pluginSetupProcessor: string;
-  multisigRepo: string;
-  adminRepo: string;
-  addresslistVotingRepo: string;
-  tokenVotingRepo: string;
-  multisigSetup: string;
-  adminSetup: string;
-  addresslistVotingSetup: string;
-  tokenVotingSetup: string;
-  ensRegistry?: string;
-};
+// Create a readonly string array from the keys of NetworkDeployment
+export const DeployedAddressesArray = [
+  "daoFactoryAddress",
+  "pluginSetupProcessorAddress",
+  "multisigRepoAddress",
+  "adminRepoAddress",
+  "addresslistVotingRepoAddress",
+  "tokenVotingRepoAddress",
+  "multisigSetupAddress",
+  "adminSetupAddress",
+  "addresslistVotingSetupAddress",
+  "tokenVotingSetupAddress",
+  "ensRegistryAddress",
+] as const;
+// export the type from the readonly string array
+export type DeployedAddresses = typeof DeployedAddressesArray[number];
+// Override helper type
+type Override<T, U> = Omit<T, keyof U> & U;
+export type NetworkDeployment = Override<
+  {
+    [address in DeployedAddresses]: string;
+  },
+  { ensRegistryAddress?: string }
+>;
 // Context input parameters
 
-export type Web3ContextParams = {
-  /** Defaults to mainnet */
-  network?: Networkish;
-  /** Required for Ethereum transactions to work */
-  signer?: Signer;
-  /** Required for Ethereum connectivity to work */
-  web3Providers?: string | JsonRpcProvider | (string | JsonRpcProvider)[];
-  /** If not provided uses default from LIVE_CONTRACTS in the provided network */
-  daoFactoryAddress?: string;
-  /** If not provided uses default from LIVE_CONTRACTS in the provided network */
-  ensRegistryAddress?: string;
-  /** If not provided uses default value */
-  gasFeeEstimationFactor?: number;
-};
+export type Web3ContextParams =
+  /** If any contract is not provided it will use the default from LIVE_CONTRACTS in the provided network */
+  & {
+    [address in DeployedAddresses]?: string;
+  }
+  & {
+    /** Defaults to mainnet */
+    network?: Networkish;
+    /** Required for Ethereum transactions to work */
+    signer?: Signer;
+    /** Required for Ethereum connectivity to work */
+    web3Providers?: string | JsonRpcProvider | (string | JsonRpcProvider)[];
+    gasFeeEstimationFactor?: number;
+  };
 export type IpfsContextParams = {
   /** If not provided uses default value */
   ipfsNodes?: { url: string; headers?: Record<string, string> }[];
@@ -41,16 +52,19 @@ export type GraphQLContextParams = {
   graphqlNodes?: { url: string }[];
 };
 
-export type Web3ContextState = {
-  network: Network;
-  signer: Signer;
-  // always going to exist because i'll default to LIVE_CONTRACTS
-  daoFactoryAddress: string;
-  // may exist depending on the network
-  ensRegistryAddress?: string;
-  web3Providers: JsonRpcProvider[];
-  gasFeeEstimationFactor: number;
-};
+export type Web3ContextState =
+  & Override<
+    {
+      [address in DeployedAddresses]: string;
+    },
+    { ensRegistryAddress?: string }
+  >
+  & {
+    network: Network;
+    signer: Signer;
+    web3Providers: JsonRpcProvider[];
+    gasFeeEstimationFactor: number;
+  };
 export type IpfsContextState = {
   ipfs: IpfsClient[];
 };