+++ modules/client/test/integration/multisig-client/estimation.test.ts
@@ -13,12 +13,7 @@ import { contextParamsLocalChain, TEST_WALLET_ADDRESS } from "../constants";
 import * as ganacheSetup from "../../helpers/ganache-setup";
 import * as deployContracts from "../../helpers/deployContracts";
 import { Server } from "ganache";
-import { Context, SupportedNetworksArray } from "@aragon/sdk-client-common";
-
-jest.spyOn(SupportedNetworksArray, "includes").mockReturnValue(true);
-jest.spyOn(Context.prototype, "network", "get").mockReturnValue(
-  { chainId: 5, name: "goerli" },
-);
+import { Context } from "@aragon/sdk-client-common";
 
 describe("Client Multisig", () => {
   describe("Estimation module", () => {