+++ modules/client/test/integration/multisig-client/methods.test.ts
@@ -47,18 +47,11 @@ import { QueryMultisigMembers } from "../../../src/multisig/internal/graphql-que
 import { SubgraphMultisigProposal } from "../../../src/multisig/internal/types";
 import {
   Context,
-  LIVE_CONTRACTS,
   PrepareInstallationStep,
   ProposalMetadata,
   ProposalStatus,
   SortDirection,
-  SupportedNetworksArray,
 } from "@aragon/sdk-client-common";
-
-jest.spyOn(SupportedNetworksArray, "includes").mockReturnValue(true);
-jest.spyOn(Context.prototype, "network", "get").mockReturnValue(
-  { chainId: 5, name: "goerli" },
-);
 
 describe("Client Multisig", () => {
   let deployment: deployContracts.Deployment;
@@ -70,6 +63,22 @@ describe("Client Multisig", () => {
     server = await ganacheSetup.start();
     deployment = await deployContracts.deploy();
     contextParamsLocalChain.daoFactoryAddress = deployment.daoFactory.address;
+    contextParamsLocalChain.pluginSetupProcessorAddress =
+      deployment.pluginSetupProcessor.address;
+    contextParamsLocalChain.multisigRepoAddress =
+      deployment.multisigRepo.address;
+    contextParamsLocalChain.adminRepoAddress = "";
+    contextParamsLocalChain.addresslistVotingRepoAddress =
+      deployment.addresslistVotingRepo.address;
+    contextParamsLocalChain.tokenVotingRepoAddress =
+      deployment.tokenVotingRepo.address;
+    contextParamsLocalChain.multisigSetupAddress =
+      deployment.multisigPluginSetup.address;
+    contextParamsLocalChain.adminSetupAddress = "";
+    contextParamsLocalChain.addresslistVotingSetupAddress =
+      deployment.addresslistVotingPluginSetup.address;
+    contextParamsLocalChain.tokenVotingSetupAddress =
+      deployment.tokenVotingPluginSetup.address;
     contextParamsLocalChain.ensRegistryAddress = deployment.ensRegistry.address;
     repoAddr = deployment.multisigRepo.address;
 
@@ -82,21 +91,6 @@ describe("Client Multisig", () => {
         contextParamsLocalChain.web3Providers as any,
       );
     }
-    LIVE_CONTRACTS.goerli.daoFactory = deployment.daoFactory.address;
-    LIVE_CONTRACTS.goerli.pluginSetupProcessor =
-      deployment.pluginSetupProcessor.address;
-    LIVE_CONTRACTS.goerli.multisigRepo = deployment.multisigRepo.address;
-    LIVE_CONTRACTS.goerli.adminRepo = "";
-    LIVE_CONTRACTS.goerli.addresslistVotingRepo =
-      deployment.addresslistVotingRepo.address;
-    LIVE_CONTRACTS.goerli.tokenVotingRepo = deployment.tokenVotingRepo.address;
-    LIVE_CONTRACTS.goerli.multisigSetup =
-      deployment.multisigPluginSetup.address;
-    LIVE_CONTRACTS.goerli.adminSetup = "";
-    LIVE_CONTRACTS.goerli.addresslistVotingSetup =
-      deployment.addresslistVotingPluginSetup.address;
-    LIVE_CONTRACTS.goerli.tokenVotingSetup =
-      deployment.tokenVotingPluginSetup.address;
     await restoreBlockTime(provider);
   });
 
@@ -228,15 +222,6 @@ describe("Client Multisig", () => {
         },
         daoAddressOrEns: dao,
       };
-      const networkSpy = jest.spyOn(JsonRpcProvider.prototype, "getNetwork");
-      const defaultGetNetworkImplementation = networkSpy
-        .getMockImplementation();
-      networkSpy.mockImplementation(() =>
-        Promise.resolve({
-          name: "goerli",
-          chainId: 31337,
-        })
-      );
       const steps = client.methods.prepareInstallation(installationParams);
       for await (const step of steps) {
         switch (step.key) {
@@ -270,7 +255,6 @@ describe("Client Multisig", () => {
             break;
         }
       }
-      networkSpy.mockImplementation(defaultGetNetworkImplementation);
     });
   });