+++ modules/client/test/integration/tokenVoting-client/encoding.test.ts
@@ -10,14 +10,7 @@ import {
 
 import { InvalidAddressError } from "@aragon/sdk-common";
 import { ADDRESS_ONE, contextParamsLocalChain } from "../constants";
-import { Context, SupportedNetworksArray } from "@aragon/sdk-client-common";
-
-const NETWORK_NAME = "goerli";
-
-jest.spyOn(SupportedNetworksArray, "includes").mockReturnValue(true);
-jest.spyOn(Context.prototype, "network", "get").mockReturnValue(
-  { chainId: 5, name: NETWORK_NAME },
-);
+import { Context } from "@aragon/sdk-client-common";
 
 describe("Token Voting Client", () => {
   beforeAll(() => {
@@ -41,7 +34,7 @@ describe("Token Voting Client", () => {
       const tokenVotingInstallPluginItem = TokenVotingClient.encoding
         .getPluginInstallItem(
           initParams,
-          NETWORK_NAME
+          "local",
         );
 
       expect(typeof tokenVotingInstallPluginItem).toBe("object");