+++ modules/client/test/integration/tokenVoting-client/estimation.test.ts
@@ -19,12 +19,7 @@ import {
 import * as ganacheSetup from "../../helpers/ganache-setup";
 import * as deployContracts from "../../helpers/deployContracts";
 import { Server } from "ganache";
-import { Context, SupportedNetworksArray } from "@aragon/sdk-client-common";
-
-jest.spyOn(SupportedNetworksArray, "includes").mockReturnValue(true);
-jest.spyOn(Context.prototype, "network", "get").mockReturnValue(
-  { chainId: 5, name: "goerli" },
-);
+import { Context } from "@aragon/sdk-client-common";
 
 describe("Token Voting Client", () => {
   describe("Estimation Module", () => {