+++ modules/client-common/CHANGELOG.md
@@ -17,6 +17,12 @@ TEMPLATE:
 -->
 
 ## [UPCOMING]
+### Changed
+-  Renamed `LIVE_CONTRACTS` properties
+
+### Added
+- Support for local chains
+## 1.4.0-rc0
 ### Added
 - Support for baseMainnet network
 ## 1.3.0-rc0